can now read decision events
moved event classes to their own file
can now read decision events
changed how the next file is handled (it's now an event instead of just changing the string of the next event)
