+++ Assets/eventManager.cs
@@ -4,390 +4,6 @@ using UnityEngine;
 using System.IO;
 
 
-public struct Decision
-{
-    public List<BaseEvent> events;
-    public string message;
-
-    public Decision(string mes, List<BaseEvent> newEvents)
-    {
-        events = newEvents;
-        message = mes;
-    }
-}
-
-public enum Effect
-{
-    gotoEvent,
-    skipEvent,
-    addDialog
-}
-
-// Base event class
-public class BaseEvent
-{
-
-    // List of potential events to go to next
-    public List<string> nextEventList;
-
-    // bool to check if event has finished
-    public bool End;
-
-
-    // constructor/destructor
-    public BaseEvent()
-    {
-
-    }
-
-    ~BaseEvent()
-    {
-
-    }
-
-
-    // virtual methods for inheritance
-    //
-    // begin is called when event starts
-    public virtual void begin()
-    {
-
-    }
-
-    // OnMouse Down is called when left mouse button is pressed
-    public virtual void OnInput()
-    {
-
-    }
-
-    // returns next event
-    public virtual string nextEvent()
-    {
-        return nextEventList[0];
-    }
-
-}
-
-// Audio event class
-public class AudioEvent : BaseEvent
-{
-
-
-    private string file;
-    private bool blocking;
-
-    private GameObject Audio;
-
-    // constructor/destructor
-    public AudioEvent(string filePath, bool block = false)
-    {
-
-        // set attributes
-
-        file = filePath;
-        blocking = block;
-
-        Audio = GameObject.FindGameObjectWithTag("AudioHandler");
-
-        End = false;
-    }
-
-    ~AudioEvent()
-    {
-
-    }
-
-    // begin is called when event starts
-    public override void begin()
-    {
-
-        // play audio from file
-
-
-        // check if event will block next event
-        if (!blocking)
-        {
-
-            // if not blocking end the event
-            End = true;
-        }
-    }
-
-    // OnMouse Down is called when left mouse button is pressed
-    public override void OnInput()
-    {
-        if (Input.GetMouseButtonDown(0))
-        {
-            // end the event
-            End = true;
-        }
-    }
-
-    // returns next event
-    public override string nextEvent()
-    {
-        return nextEventList[0];
-    }
-
-}
-
-public class DecisionEvent : BaseEvent
-{
-
-    // attributes
-    private int currentSelect;
-    private List<Decision> selection;
-
-    GameObject UI;
-    GameObject EManager;
-
-    public DecisionEvent(List<Decision> newSelection)
-    {
-        selection = newSelection;
-
-        UI = GameObject.FindGameObjectWithTag("Canvas");
-        EManager = GameObject.FindGameObjectWithTag("EventManager");
-
-        End = false;
-    }
-
-    ~DecisionEvent()
-    {
-
-    }
-
-    public override void begin()
-    {
-        currentSelect = 0;
-        // display the options
-        //UI.GetComponent<UIhandler>().displaySelection(selection);
-        //UI.GetComponent<UIhandler>().displayCursor(currentSelect);
-    }
-
-    public override void OnInput()
-    {
-
-        if (Input.GetKeyDown(KeyCode.DownArrow))
-        {
-            currentSelect++;
-            currentSelect %= selection.Count;
-
-            //UI.GetComponent<UIhandler>().displayCursor(currentSelect);
-        }
-
-        if (Input.GetKeyDown(KeyCode.Return))
-        {
-            Choose();
-        }
-    }
-
-    private void Choose()
-    {
-
-        // load events
-        // add events to events list
-        // end event
-        /*
-        switch (selection[currentSelect].type)
-        {
-            case Effect.gotoEvent:
-                gotoEvent();
-                break;
-            case Effect.skipEvent:
-                skipEvent();
-                break;
-            case Effect.addDialog:
-                addDialog();
-                break;
-        }
-        */
-    }
-
-    private void addDialog()
-    {
-
-    }
-
-    private void skipEvent()
-    {
-        //EManager.GetComponent<eventManager>().GotoNextEvent();
-    }
-
-    private void gotoEvent()
-    {
-        //EManager.GetComponent<eventManager>().setNextEvent(selection[currentSelect].data);
-        End = true;
-    }
-
-    public override string nextEvent()
-    {
-
-        return "";
-    }
-
-}
-
-
-// Character Change event class
-public class CharacterChangeEvent : BaseEvent
-{
-
-    // attributes
-    private string file;
-    private string charNum;
-    private bool blocking;
-
-    private GameObject UI;
-
-    // constructor/destructor
-    public CharacterChangeEvent(string filePath, string characterNum, bool block = false)
-    {
-
-        // set attributes
-
-        file = filePath;
-        charNum = characterNum;
-        blocking = block;
-
-        UI = GameObject.FindGameObjectWithTag("Canvas");
-
-        End = false;
-    }
-
-    ~CharacterChangeEvent()
-    {
-
-    }
-
-
-    // begin is called when event starts
-    public override void begin()
-    {
-
-        // display character
-        UI.GetComponent<UIhandler>().changeImageSprite(charNum, file);
-
-
-        // check if event is blocking
-        if (!blocking)
-        {
-
-            // end the event
-            End = true;
-        }
-
-    }
-
-    // OnMouse Down is called when left mouse button is pressed
-    public override void OnInput()
-    {
-        if (Input.GetMouseButtonDown(0))
-        {
-
-            // end the event
-            End = true;
-        }
-    }
-
-    // returns next event
-    public override string nextEvent()
-    {
-        return nextEventList[0];
-    }
-
-}
-
-
-// dialog event class
-public class DialogEvent : BaseEvent
-{
-
-    // List of dialog used
-    public List<Dialog> dialog;
-
-    // index of current dialog
-    private int currentDialog;
-
-    // pointers to UI/AUDIO objects
-    private GameObject UI;
-    private GameObject Audio;
-
-
-
-    // constructor/destructor
-    public DialogEvent(List<Dialog> newDialog)//, string nextEvent, GameObject setUI, GameObject setAudio )
-    {
-
-        // set current dialog to start
-        currentDialog = 0;
-
-        // set attributes
-        dialog = newDialog;
-
-        //nextEventList.Add(nextEvent);
-
-        UI = GameObject.FindGameObjectWithTag("Canvas");
-        Audio = GameObject.FindGameObjectWithTag("Canvas");
-
-        //UI = setUI;
-        //Audio = setAudio;
-
-        End = false;
-
-    }
-
-
-    // called when event starts
-    public override void begin()
-    {
-        Debug.Log("begin is happening");
-        UI.GetComponent<UIhandler>().changeText(dialog[currentDialog]);
-        Debug.Log(dialog[currentDialog].name);
-        // UI display first dialog
-
-    }
-
-    // when mouse button down
-    public override void OnInput()
-    {
-
-        if (Input.GetMouseButtonDown(0))
-        {
-            Debug.Log("displaying next text");
-            Debug.Log(dialog[currentDialog].message);
-            Debug.Log(dialog[currentDialog].name);
-
-            // goto next dialog 
-            currentDialog++;
-
-            // check if at end of dialog
-            if (currentDialog < dialog.Count)
-            {
-
-                // if there is dialog display it
-
-                //Debug.Log("mouse down displaying new poop hehe got 'em");
-                UI.GetComponent<UIhandler>().changeText(dialog[currentDialog]);
-                Debug.Log(dialog[currentDialog].name);
-                // UI display dialog[currentDialog]
-            }
-            else
-            {
-
-                // else end the event
-                End = true;
-            }
-        }
-    }
-
-    // returns location of next event
-    public override string nextEvent()
-    {
-
-        return nextEventList[0];
-    }
-
-}
 
 public class eventManager : MonoBehaviour
 {
@@ -396,6 +12,8 @@ public class eventManager : MonoBehaviour
     static char ENDEVENT = '~';
     static char KEYSTART = '[';
     static char KEYEND = ']';
+    static char DECISIONSTART = '{';
+    static char DECISIONEND = '}';
 
 
     
@@ -496,7 +114,7 @@ public class eventManager : MonoBehaviour
     // create event types
     //
     // adds Dialog event to event list
-    private void CreateDialogEvent(StreamReader reader)
+    private DialogEvent CreateDialogEvent(StreamReader reader)
     {
 
         // create variables for creating dialog event
@@ -621,24 +239,138 @@ public class eventManager : MonoBehaviour
 
             }
         }
-        
-      
 
 
+
+        return new DialogEvent(DialogList);
         // create and add dialog event to the event list
-        eventList.Add(new DialogEvent(DialogList));
+        
 
     }
 
 
     // adds decision event to event list
-    private void CreateDecisionEvent(StreamReader reader)
+    private DecisionEvent CreateDecisionEvent(StreamReader reader)
     {
+
+        List<Decision> DecisionList = new List<Decision>();
+        List<BaseEvent> eventList = new List<BaseEvent>();
+
+        char character;
+        string message = "";
+        string buffer = "";
+
+        while (reader.Peek() > -1)
+        {
+
+            buffer = "";
+            message = "";
+
+            character = skipWhiteSpace(reader);
+
+            //Debug.Log(character);
+
+            if (character == DECISIONSTART)
+            {
+
+                Debug.Log("decision started");
+
+                while (reader.Peek() > -1)
+                {
+
+                    buffer = "";
+
+                    character = skipWhiteSpace(reader);
+
+                    //Debug.Log(character);
+
+                    if (character == DECISIONEND)
+                    {
+                        //Debug.Log("decision has ended");
+                        break;
+                    }
+
+                    else if (character == KEYSTART)
+                    {
+
+                        while (reader.Peek() > -1)
+                        {
+
+                            character = (char)reader.Read();
+
+                            if (character == KEYEND)
+                            {
+
+                                break;
+                            }
+                            else
+                            {
+                                buffer += character;
+                            }
+                        }
+
+                        switch (buffer.ToUpper())
+                        {
+                            case "MESSAGE":
+                                Debug.Log("adding message");
+                                readLine(reader, ref message, true);
+                                break;
+                            case "DIALOG":   // if event type dialog
+                                Debug.Log("dialogeventpoo");
+                                eventList.Add( CreateDialogEvent(reader));
+                                break;
+                            case "EVENTFILE":
+                                Debug.Log("eventfile");
+                                AddEventFile(reader);
+                                break;
+                            case "AUDIO":
+                                Debug.Log("audioevent");
+                                CreateAudioEvent(reader);
+                                break;
+                            case "CHARACTERLOAD":
+                                Debug.Log("characterLoadEvent");
+                                CreateCharacterChangeEvent(reader);
+                                break;
+                            case "END":   // if end of file
+                                //Debug.Log("end of file");
+                                //endOfFile = true;
+                                break;
+                            default:
+                                break;
+                        }
+                        //Debug.Log("loaded event in decision");
+                    }
+                    
+                }
+
+                //Debug.Log("decision being added");
+                DecisionList.Add(new Decision(message, eventList));
+
+                eventList = new List<BaseEvent>();
+                message = "";
+            }
+            else if (character == ENDEVENT)
+            {
+                break;
+                
+            }
+            else
+            {
+                //  parsing error
+            }
+
+            
+
+        }
+
+        Debug.Log("decision count");
+        Debug.Log(DecisionList.Count);
+        return new DecisionEvent(DecisionList);
 
     }
 
     // adds audio event to event list
-    private void CreateAudioEvent(StreamReader reader)
+    private AudioEvent CreateAudioEvent(StreamReader reader)
     {
 
         // create variables for making audio event
@@ -738,7 +470,7 @@ public class eventManager : MonoBehaviour
 
 
         // add new Audio event to list
-        eventList.Add(new AudioEvent(filePath, blocking));
+        return new AudioEvent(filePath, blocking);
 
     }
 
@@ -773,7 +505,7 @@ public class eventManager : MonoBehaviour
 
 
     // adds Character Change event to event list
-    private void CreateCharacterChangeEvent(StreamReader reader)
+    private CharacterChangeEvent CreateCharacterChangeEvent(StreamReader reader)
     {
 
         // create variables for making event
@@ -902,14 +634,14 @@ public class eventManager : MonoBehaviour
         }
 
         // add character change event to event list
-        eventList.Add(new CharacterChangeEvent(filePath, charNum, blocking));
+        return new CharacterChangeEvent(filePath, charNum, blocking);
 
     }
 
     
 
     // adds event file to next events list
-    private void AddEventFile(StreamReader reader)
+    private SetNextFileEvent AddEventFile(StreamReader reader)
     {
 
         // set file directory
@@ -947,7 +679,7 @@ public class eventManager : MonoBehaviour
 
 
         // add eventfile to list
-        nextEvent = file;
+        return new SetNextFileEvent(file);
     }
 
     private void GotoNextEvent()
@@ -964,6 +696,7 @@ public class eventManager : MonoBehaviour
             currentEvent = eventList[eventIndex];
 
             // begin the event
+            Debug.Log("beginning an event");
             currentEvent.begin();
         }
         else
@@ -977,6 +710,18 @@ public class eventManager : MonoBehaviour
         
     }
 
+    public void InsertEvents(List<BaseEvent> newEvents)
+    {
+
+        eventList.InsertRange(eventIndex + 1, newEvents);
+        
+    }
+
+    public void SetNextEvent(string path)
+    {
+
+        nextEvent = path;
+    }
 
     // load new events
     private bool LoadEvent()
@@ -1041,22 +786,25 @@ public class eventManager : MonoBehaviour
             // check type of event
             switch (buffer.ToUpper())
             {
-
+                case "DECISION":
+                    Debug.Log("decisionEvent");
+                    eventList.Add(CreateDecisionEvent(reader));
+                    break;
                 case "DIALOG":   // if event type dialog
                     Debug.Log("dialogevent");
-                    CreateDialogEvent(reader);
+                    eventList.Add( CreateDialogEvent(reader) );
                     break;
                 case "EVENTFILE":
                     Debug.Log("eventfile");
-                    AddEventFile(reader);
+                    eventList.Add( AddEventFile(reader) );
                     break;
                 case "AUDIO":
                     Debug.Log("audioevent");
-                    CreateAudioEvent(reader);
+                    eventList.Add( CreateAudioEvent(reader) );
                     break;
                 case "CHARACTERLOAD":
                     Debug.Log("characterLoadEvent");
-                    CreateCharacterChangeEvent(reader);
+                    eventList.Add( CreateCharacterChangeEvent(reader) );
                     break;
                 case "END":   // if end of file
                     Debug.Log("end of file");
